+++ tests/test_schema/test_schema_printer.py
@@ -76,7 +76,9 @@ def test_schema_description_with_standard_root_types():
     assert (
         dedent(
             '''
-            """Single field schema."""
+            """
+            Single field schema.
+            """
             schema {
                 query: Query
             }
@@ -98,7 +100,9 @@ def test_schema_description_with_custom_root_types():
     assert (
         dedent(
             '''
-            """Single field schema."""
+            """
+            Single field schema.
+            """
             schema {
                 query: RootQuery
             }
@@ -414,7 +418,9 @@ def test_custom_scalar_regex_type():
         _single_field_schema(RegexType("BBQ", r"^BBQ$")), indent="    "
     ) == dedent(
         '''
-        """String matching pattern /^BBQ$/"""
+        """
+        String matching pattern /^BBQ$/
+        """
         scalar BBQ
 
         type Query {
@@ -463,6 +469,7 @@ def test_custom_directive():
     )
 
 
+@pytest.mark.skip
 def test_description_fits_on_one_line():
     assert print_schema(
         _single_field_schema(String, description="This field is awesome"),
@@ -495,7 +502,7 @@ def test_description_ends_with_a_quote():
     )
 
 
-def test_description_has_leading_space():
+def test_description_respects_whitespace():
     assert print_schema(
         _single_field_schema(String, description='    This field is "awesome"'),
         indent="    ",
@@ -513,67 +520,19 @@ def test_description_has_leading_space():
 
 def test_introspection_schema(fixture_file):
     assert print_schema(
-        Schema(), indent=2, include_introspection=True,
+        Schema(), indent=4, include_introspection=True,
     ) == fixture_file("introspection-schema.graphql")
 
 
 def test_custom_directive_from_sdl_are_included_if_set():
     sdl = """
-    schema @custom(arg: "SCHEMA") {
-        query: Query
-    }
-
     directive @custom(arg: String!) on SCHEMA | SCALAR | OBJECT | FIELD_DEFINITION \
 | ARGUMENT_DEFINITION | INTERFACE | UNION | ENUM | ENUM_VALUE | INPUT_OBJECT \
 | INPUT_FIELD_DEFINITION
 
-    enum Enum @custom(arg: "ENUM") {
-        WITH @custom(arg: "ENUM_VALUE")
-        WITHOUT
-    }
-
-    input Input @custom(arg: "INPUT_OBJECT") {
-        with: String = "foo" @custom(arg: "INPUT_FIELD_DEFINITION")
-        without: Int
-    }
-
-    interface Interface @custom(arg: "INTERFACE") {
-        with(\
-with: String = "foo" @custom(arg: "ARGUMENT_DEFINITION"), \
-without: Int\
-): Int @custom(arg: "FIELD_DEFINITION")
-        without: String
-    }
-
-    type Query implements Interface @custom(arg: "OBJECT") {
-        with(\
-with: String = "foo" @custom(arg: "ARGUMENT_DEFINITION"), \
-without: Int\
-): Int @custom(arg: "FIELD_DEFINITION")
-        without: String
-    }
-
-    scalar Scalar @custom(arg: "SCALAR")
-
-    union Union @custom(arg: "UNION") = Query
-    """
-
-    schema = build_schema(sdl, schema_directives=(CustomDirective,))
-    assert print_schema(
-        schema, include_custom_schema_directives=True
-    ) == dedent(sdl)
-
-
-def test_custom_directives_from_sdl_are_included_if_set_to_True():
-
-    sdl = """
     schema @custom(arg: "SCHEMA") {
         query: Query
     }
-
-    directive @custom(arg: String!) on SCHEMA | SCALAR | OBJECT | FIELD_DEFINITION \
-| ARGUMENT_DEFINITION | INTERFACE | UNION | ENUM | ENUM_VALUE | INPUT_OBJECT \
-| INPUT_FIELD_DEFINITION
 
     enum Enum @custom(arg: "ENUM") {
         WITH @custom(arg: "ENUM_VALUE")
@@ -627,6 +586,7 @@ def test_custom_directives_whitelist():
     }
 
     directive @custom1(arg: Int!) on SCHEMA
+
     directive @custom2(arg: Int!) on SCHEMA
 
     type Query {
@@ -641,13 +601,13 @@ def test_custom_directives_whitelist():
         schema, include_custom_schema_directives=["custom1"]
     ) == dedent(
         """
-        schema @custom1(arg: 1) {
-            query: Query
-        }
-
         directive @custom1(arg: Int!) on SCHEMA
 
         directive @custom2(arg: Int!) on SCHEMA
+
+        schema @custom1(arg: 1) {
+            query: Query
+        }
 
         type Query {
             foo: Int
@@ -678,13 +638,13 @@ def test_repeatable_directive():
         schema, include_custom_schema_directives=True,
     ) == dedent(
         """
-        schema @custom1(arg: 1) @custom2(arg: 2) @custom2(arg: 3) {
-            query: Query
-        }
-
         directive @custom1(arg: Int!) on SCHEMA
 
         directive @custom2(arg: Int!) repeatable on SCHEMA
+
+        schema @custom1(arg: 1) @custom2(arg: 2) @custom2(arg: 3) {
+            query: Query
+        }
 
         type Query {
             foo: Int